fix template

+++ app/logic/Admin_user_logic.php
@@ -29,16 +29,23 @@ class Admin_user_logic extends Basetool
          
          $_this = new self();
 
-         $result = array(
-                        "account"       => isset($data["account"]) ? trim($data["account"]) : "",
-                        "password"      => isset($data["password"]) ? bcrypt($_this->strFilter($data["password"])) : "",
-                        "real_name"     => isset($data["real_name"]) ? $_this->strFilter($data["real_name"]) : "",
-                        "mobile"        => isset($data["mobile"]) ? $_this->strFilter($data["mobile"]) : "",
-                        "parents_id"    => isset($data["parents_id"]) ? intval($data["parents_id"]) : 0,
-                        "status"        => isset($data["active"]) ? intval($data["active"]) : 0,
-                        "created_at"    => date("Y-m-d H:i:s"),
-                        "updated_at"    => date("Y-m-d H:i:s")
-                     );
+         $result = array();
+
+         if ( !empty($data) && is_array($data) ) 
+         {
+
+            $result = array(
+                           "account"       => isset($data["account"]) ? trim($data["account"]) : "",
+                           "password"      => isset($data["password"]) ? bcrypt($_this->strFilter($data["password"])) : "",
+                           "real_name"     => isset($data["real_name"]) ? $_this->strFilter($data["real_name"]) : "",
+                           "mobile"        => isset($data["mobile"]) ? $_this->strFilter($data["mobile"]) : "",
+                           "parents_id"    => isset($data["parents_id"]) ? intval($data["parents_id"]) : 0,
+                           "status"        => isset($data["active"]) ? intval($data["active"]) : 0,
+                           "created_at"    => date("Y-m-d H:i:s"),
+                           "updated_at"    => date("Y-m-d H:i:s")
+                        );
+
+         }
 
          return $result;
 
@@ -49,18 +56,24 @@ class Admin_user_logic extends Basetool
 
          $_this = new self();
 
-         $result = array(
-                        "real_name"     => isset($data["real_name"]) ? $_this->strFilter($data["real_name"]) : "",
-                        "mobile"        => isset($data["mobile"]) ? $_this->strFilter($data["mobile"]) : "",
-                        "status"        => isset($data["active"]) ? intval($data["active"]) : 0,
-                        "updated_at"    => date("Y-m-d H:i:s")
-                     );
+         $result = array();
 
-         if (!empty($data["password"])) 
+         if ( !empty($data) && is_array($data) ) 
          {
-            $result["password"] = bcrypt($_this->strFilter($data["password"]));
-         }
 
+            $result = array(
+                           "real_name"     => isset($data["real_name"]) ? $_this->strFilter($data["real_name"]) : "",
+                           "mobile"        => isset($data["mobile"]) ? $_this->strFilter($data["mobile"]) : "",
+                           "status"        => isset($data["active"]) ? intval($data["active"]) : 0,
+                           "updated_at"    => date("Y-m-d H:i:s")
+                        );
+
+            if (!empty($data["password"])) 
+            {
+               $result["password"] = bcrypt($_this->strFilter($data["password"]));
+            }
+
+         }
 
          return $result;
 
@@ -71,17 +84,30 @@ class Admin_user_logic extends Basetool
 
          $auth = array();
 
-         foreach ($rel_data as $row)
+         $result = array();
+
+         if ( !empty($data) && !empty($rel_data) ) 
          {
-            $auth[$row->user_id][] = $row->name;
+
+            foreach ($rel_data as $row)
+            {
+
+               $auth[$row->user_id][] = $row->name;
+            
+            }
+
+            foreach ($data as &$row) 
+            {
+            
+               $row->auth = isset($auth[$row->id]) ? $auth[$row->id] : array() ;
+            
+            }
+
+            $result = $data;
+
          }
 
-         foreach ($data as &$row) 
-         {
-            $row->auth = isset($auth[$row->id]) ? $auth[$row->id] : array() ;
-         }
-
-         return $data;
+         return $result;
 
    }  
 
@@ -90,12 +116,17 @@ class Admin_user_logic extends Basetool
 
          $result = array();
 
-         foreach ($data as $key => $value)
+         if ( !empty($data) && is_array($data) ) 
          {
-            $result[] = array(
-                              "user_id"   => intval($user_id),
-                              "role_id"   => intval($value)
-                        );
+
+            foreach ($data as $key => $value)
+            {
+               $result[] = array(
+                                 "user_id"   => intval($user_id),
+                                 "role_id"   => intval($value)
+                           );
+            }
+
          }
 
          return $result;
@@ -105,7 +136,16 @@ class Admin_user_logic extends Basetool
    public static function get_user( $id = 0 )
    {
 
-         return Admin_user::get_user( $id );
+         $result = false;
+
+         if ( !empty($id) && is_int($id) ) 
+         {
+
+            $result = Admin_user::get_user( $id );
+
+         }
+
+         return $result;
 
    }
 
@@ -167,42 +207,102 @@ class Admin_user_logic extends Basetool
    public static function get_user_role_by_id( $id )
    {
 
-         return Admin_user::get_user_role_by_id( $id );
-         
+         $result = array();
+
+         if ( !empty($id) && is_int($id) ) 
+         {
+
+            $result = Admin_user::get_user_role_by_id( $id );
+
+         }
+
+         return $result;
+
    }
 
    public static function add_user( $data )
    {
 
-         return Admin_user::add_user( $data );
+         $result = false;
+
+         if ( !empty($data) && is_array($data) ) 
+         {
+
+            $result = Admin_user::add_user( $data );
+
+         }
+
+         return $result ;
          
    }
 
    public static function edit_user( $data, $user_id )
    {
 
-         Admin_user::edit_user( $data, $user_id );
+         $result = false;
+
+         if ( !empty($data) && is_array($data) && !empty($user_id) && is_int($user_id) ) 
+         {
+
+            Admin_user::edit_user( $data, $user_id );
+
+            $result = true;
+
+         }
+
+         return $result;
          
    }
 
    public static function add_user_role( $data )
    {
 
-         Admin_user::add_user_role( $data );
-         
+         $result = false;
+
+         if ( !empty($data) && is_array($data) ) 
+         {
+
+             Admin_user::add_user_role( $data );
+
+            $result = true;
+
+         }
+
+         return $result;
+
    }
 
    public static function delete_user_role( $user_id )
    {
 
-         Admin_user::delete_user_role( $user_id );
+         $result = false;
+
+         if ( !empty($user_id) && is_int($user_id) ) 
+         {
+
+            Admin_user::delete_user_role( $user_id );
+
+            $result = true;
+
+         }
+
+         return $result;
          
    }
 
    public static function get_user_id( $data )
    {
 
-         return Admin_user::get_user_id( $data );
+         $result = array();
+
+         if ( !empty($data) && is_array($data) ) 
+         {
+
+            $result = Admin_user::get_user_id( $data );
+
+         }
+
+         return $result;
          
    }
 
@@ -211,11 +311,28 @@ class Admin_user_logic extends Basetool
 
          $result = array();
 
-         $data = Admin_user::get_user_id_by_role( $role_id );
-
-         foreach ($data as $row) 
+         if ( !empty($role_id) && is_int($role_id) ) 
          {
-            $result[] = $row->user_id;
+
+            $data = Admin_user::get_user_id_by_role( $role_id );
+
+            if ( !empty($data) ) 
+            {
+
+               foreach ($data as $row) 
+               {
+
+                  if ( is_object($row) ) 
+                  {
+
+                     $result[] = $row->user_id;
+
+                  }    
+               
+               }
+
+            }
+
          }
 
          return $result;
@@ -226,6 +343,8 @@ class Admin_user_logic extends Basetool
    {
 
          $_this = new self();
+
+         $len = is_int($len) ? $len : 3 ;
 
          $string = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
 
@@ -245,15 +364,23 @@ class Admin_user_logic extends Basetool
 
          }
 
-
-        return $result;
+         return $result;
 
    }
 
    public static function check_store_code_repeat( $store_code )
    {
 
-         return Admin_user::check_store_code_repeat( $store_code );
+         $result = 0;
+
+         if ( !empty($store_code) && is_string($store_code) ) 
+         {
+
+            $result = Admin_user::check_store_code_repeat( $store_code );
+
+         }
+
+         return $result;
 
    }
 
@@ -295,12 +422,13 @@ class Admin_user_logic extends Basetool
                   );
 
          return $result;     
+
    }
 
    public static function is_admin( $data )
    {
 
-         $result = $data["user_id"] == 1 ? true :false;
+         $result = !empty($data) && is_array($data) && $data["user_id"] == 1 ? true : false;
 
          return $result;
 
@@ -309,9 +437,16 @@ class Admin_user_logic extends Basetool
    public static function is_sub_admin( $data )
    {
 
-         $data = Admin_user::is_sub_admin( $data["user_id"] );
+         $result = false;
 
-         $result = !empty($data) && $data->parents_id > 0 ? false : true; 
+         if ( !empty($data) && is_array($data) )
+         {
+
+            $data = Admin_user::is_sub_admin( $data["user_id"] );
+
+            $result = !empty($data) && $data->parents_id > 0 ? false : true; 
+
+         } 
 
          return $result;   
 
@@ -320,16 +455,34 @@ class Admin_user_logic extends Basetool
    public static function get_parents_id( $data )
    {
 
-         $data = Admin_user::is_sub_admin( $data["user_id"] );
+         $result = 0 ;
 
-         return $data->parents_id;   
+         if ( !empty($data) && is_array($data) )
+         {
+
+            $data = Admin_user::is_sub_admin( $data["user_id"] );
+
+            $result = is_object($data) ? $data->parents_id : 0 ; 
+
+         } 
+
+         return $result;   
 
    }
 
    public static function insert_store( $data )
    {
 
-         return Admin_user::insert_store( $data );
+         $result = false;
+
+         if ( !empty($data) && is_array($data) ) 
+         {
+
+            $result = Admin_user::insert_store( $data );
+
+         }
+
+         return $result;
          
    }
 
@@ -345,9 +498,9 @@ class Admin_user_logic extends Basetool
 
             $result = Admin_user::get_user_photo( $Login_user["user_id"] );
 
-         }
+            $result->photo = !empty($result->photo) ? $result->photo : "_images/user_default.png" ;
 
-         $result->photo = !empty($result->photo) ? $result->photo : "_images/user_default.png" ;
+         }
 
          return $result;
          
@@ -358,18 +511,25 @@ class Admin_user_logic extends Basetool
 
          $_this = new self();
 
-         $Login_user = Session::get( 'Login_user' );
+         $result = false;
 
-         $ori_image = $_this->get_user_image( $Login_user["user_id"] );
-
-         if ( !empty($ori_image->photo) && file_exists( $ori_image->photo ) ) 
+         if ( !empty($photo_upload_files) ) 
          {
 
-            unlink( $ori_image->photo );
+            $Login_user = Session::get( 'Login_user' );
+
+            $ori_image = $_this->get_user_image( (int)$Login_user["user_id"] );
+
+            if ( !empty($ori_image->photo) && file_exists( $ori_image->photo ) ) 
+            {
+
+               unlink( $ori_image->photo );
+
+            }
+
+            $result = Admin_user::edit_user_photo( $photo_upload_files, $Login_user["user_id"] );
 
          }
-
-         $result = Admin_user::edit_user_photo( $photo_upload_files, $Login_user["user_id"] );
 
          return $result;
          
@@ -380,7 +540,18 @@ class Admin_user_logic extends Basetool
    public static function extend_user_deadline( $user_id, $data )
    {
 
-      return Shop_logic::add_use_record( $user_id, $data, $type = 1 );
+      $result = false;
+
+      if ( !empty($user_id) && is_int($user_id) && !empty($data) && is_array($data) ) 
+      {
+
+         Shop_logic::add_use_record( (int)$user_id, $data, $type = 1 );
+
+         $result = true;
+
+      }
+
+      return $result;
          
    }
 
@@ -389,8 +560,18 @@ class Admin_user_logic extends Basetool
    public static function get_user_image( $user_id )
    {
 
-      return Admin_user::get_user_image( $user_id );
+      $result = false;
+
+      if ( !empty($user_id) && is_int($user_id) ) 
+      {
+
+         $result = Admin_user::get_user_image( $user_id );
+
+      }
+
+      return $result;
          
    }
+
 
 }
